Extract Shared Components from presenter
- To avoid semantically confusing inheritance a shared components
  folder has been created. It encapsulates common presenters behaviour
