+++ app/presenters/regional_strategy_template_presenter.rb
@@ -6,8 +6,4 @@ class RegionalStrategyTemplatePresenter < ArticleTemplatePresenter
   def forum_links_component
     extract_links(forum_links_block.try(:content).to_s)
   end
-
-  def strategy_text_component
-    view.strip_tags(strategy_text_block.try(:content).to_s)
-  end
 end
